Switch Literal and Variable from u64 to u32
- Change Variable(u64) to Variable(u32)
- Change Literal(u64) to Literal(u32)
- Update MAX_LITERAL to 1 << 31 (2B variables still supported)
- Update VariableRegister to use u32 internally
- Update sosorted call to use u32 slice

Benchmark results show 5-6% improvement on larger problems (dubois)
due to better cache utilization and 2x SIMD throughput.

+++ src/instance/variable.rs
@@ -2,7 +2,7 @@ use std::fmt;
 
 #[derive(Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
 #[repr(transparent)]
-pub struct Variable(pub u64);
+pub struct Variable(pub u32);
 
 impl fmt::Debug for Variable {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
